requirements.txt
appdirs==1.4.4
asgiref==3.3.4
astroid==2.5.6
certifi==2020.12.5
colorama==0.4.4
distlib==0.3.1
Django==3.2.4
djangorestframework==3.12.4
filelock==3.0.12
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
Pillow==8.3.1
pipenv==2020.11.15
pylint==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
virtualenv==20.4.6
virtualenv-clone==0.5.4
whitenoise==5.3.0
wrapt==1.12.1
